requirements.txt
asgiref==3.6.0
Django==4.1.7
environs==9.5.0
marshmallow==3.19.0
packaging==23.0
python-dotenv==1.0.0
sqlparse==0.4.3
django-phonenumber-field[phonenumberslite]
